Rename deliver_async to deliver_later
deliver_async is misleading because this function is not meant to be
awaited on. Instead it's meant as more of a "send it and forget it"
function. It can always be customized with a custom adapter if you want
to do something different. I think deliver_later is more flexible and
doesn't have as many misleading connotations.

Closes #65