Better assertion helpers for Bamboo (#109)
Instead of using the ExUnit assert_receive and refute_receive, the
Bamboo helpers now use receive directly and only show emails from the
process mailbox. It also shows helpful advice when things aren't working
as expected.

Setting a `refute_timeout` configuration option is required when
refuting that emails are sent when using the `process_name` feature.
This is because we don't want to set a time out that is too low or too
high. If it's too low the user may have delivered an email after the
timeout is sent, resulting in a false pass. And if the timeout is too
high then their tests will be unnecessarily slow. So instead we require
the timeout to be set and offer some guidance on how to set it and what
the timeout might be.

Note, that the easiest way to get around this is to try not to use the
`process_name` feature :) This can be done by avoiding Tasks and
acceptance tests. It might be better to use `deliver_later/1` and
controller tests when testing email delivery.

The assert timeout is not configurable (at least for now), because even
if the timeout is 100ms, it will pass as soon as the email is delivered.