Start TaskSupervisor by default
The reason for this is that if two applications start bamboo and both of
them added Bamboo.TaskSupervisorStrategy.child_spec to their children,
then there would be an error because it was started twice.

This circumvents that by starting it just once when Bamboo starts. This
also simplifies the installation instructions.