Remove warnings-as-errors until deps are updated
Right now CI isn't passing due to some deps that have warnings under the new version of Erlang. So I'll remove this and add it back in once we've had a chance to figure get it passing without any warnings again. For now this is making it harder for contributors to get things in so I'll remove it.