Mandrill Template Support (#176)
Adds the ability to use Mandrill's [built in template support](https://mandrillapp.com/api/docs/messages.JSON.html#method=send-template). This means template rendering can be fully offloaded to Mandrill. This is done by adding a template function to `MandrillHelper` that inserts template config into the private params. These params are detected when trying to deliver the email which switches over the API path and inserts params into the request.

Here is an example of how it can be used below. Please note the only new public api is `Bamboo.MandrillHelper.template` the rest of the config can be done using existing functions.

``` elixir
new_email
|> to("john@example.com")
|> from("no-reply@service.com")
|> Bamboo.MandrillHelper.template("invoice")
|> Bamboo.MandrillHelper.put_param("merge_language", "handlebars")
|> Bamboo.MandrillHelper.put_param("global_merge_vars", [ %{ "name": "reference_number", "content": "ABC001" }])
|> Mailer.deliver_now
```

Tests are included and this is working against a production Mandrill account. Regular message sending via the API is unaffected by this change.