Allows dynamic configurations in deliveries
At times we need custom configurations for sending specific emails. The
custom configurations might be needed because specific email addresses
need to move through specific servers. Instead of calling
`Bamboo.Mailer.deliver_*` directly and bypassing the public API this
changes the API to take in an optional second argument that is merged
into the default configuration. It will take in a map or a list of
keyword arguments.

Amos King @adkron <amos@binarynoggin.com>