Allow to use the complete API URL for the Mailgun adapter
This commit allows to use the complete API URL instead of your domain for the
Mailgun adapter.

When I was trying to use this adapter for the first time I tried to follow the
same pattern as in `curl`:

```
curl -s --user 'api:key-3ax6xnjp29jd6fds4gc373sgvjxteol0' \
https://api.mailgun.net/v3/samples.mailgun.org/messages \
 -F from='Excited User <excited@samples.mailgun.org>' \
 -F to='devs@mailgun.net' \
 -F subject='Hello' \
 -F text='Testing some Mailgun awesomeness!'
```

It was a mistake trying to do this, but the current implementation do not emit
any warning about this. So, instead of emit a warning I added support for
passing the complete API URL instead of the domain.