lazily render debug logs
Without this, `inspect()` is *always* called for every email, regardless
of the log level. After this change, `inspect()` only gets called if the
log level is debug or more verbose.

This can potentially save a massive amount of time depending on the size
of the Email struct (and whatever `:assigns`) it has.