lazily render debug logs (#438)
* lazily render debug logs

Without this, `inspect()` is *always* called for every email, regardless
of the log level. After this change, `inspect()` only gets called if the
log level is debug or more verbose.

This can potentially save a massive amount of time depending on the size
of the Email struct (and whatever `:assigns`) it has.

* change inspect limit :infinity to 150

Using infinity can lead to some extremely slow logging when large
structs are being passed in. See #438 for more details.