Add ability to bypass list management for SendGridAdapter (#458)
SendGrid offers the option to mark emails as being able to bypass list management (suppression of unsubscribed/bounced emails). This is useful for emails which are compulsory to send, regardless of marketing/personal preferences, like Terms of Service updates, or security notices. This adds a with_bypass_list_management helper to enable this.