Add .t() spec type to Bamboo.Attachment (#389)
Added (probably) missing `.t()` type spec to `Bamboo.Attachment`.

In most deps, the prepared types for dialyzer's specs makes it easier to prepare one's own specs. I decided to insert this type in the Attachment module. In one of my projects, I tried to use it, and I was surprised when dialyzer returned `type does not exist`.