Clarify changes and test phases in mailer test
The variance between tests in the mailer test case all relate to the
changes in adapters and what those adapters support. Currently each
change in adapter includes a new mailer, a new config declaration, and a
new addition to the environment. As all of these changes are isolated
near the top of the file, it is less clear when reading tests lower in
the file why the response of one test is expected to differ from that of
another; there is too much separation between test setup, execution, and
assertions.

This commit reduces the amount of setup by changing only the adapter
through the use of tags and ExUnit's setup macro. This allows tests that
pivot on a particular adapter to define the adapter and declare it's use
right next to the tests that use it, bringing setup, execution, and
assertions closer together. This commit also brings common case tests to
the top of the file and groups non-common tests with describe blocks.