Add option to deliver_now to return adapter response (#466)
Adds a `response: true` option to `deliver_now` that returns a tuple of the `Email` struct and the response from the adapter's `deliver` call. This is a non-breaking change, since without the `response: true` option everything will work as usual, and the `Email` will be returned, but this adds the option to pass that additional argument in and also get the response back.