Fix SendGridAdapter.deliver/2 Test :100:
The recent rebase with `thoughtbot/bamboo:master` introduced some new
tests to `SendGridAdapterTest`, specifically the use of a new `Reply-To`
header with a Tuple value. If this is not filtered from the headers when
`SendGridAdapter.put_headers/2` is called, then a `Jason.Encoder` error
will be raised.

The original commit for `put_headers/2` included a fix for the
`reply-to` header. This commit adds the fix for the recently introduced
`Reply-To` header test.