refactor: Improve error reporting in mailgun
Mailgun adapter was raising errors with encoded body.  This caused the
api_error handler to report one large string as the param. This was hard
to debug and know which email was throwing the error because
printable_limit config of `inspect\2` was truncating the string to 4092
characters. Dude to the way maps are sorted alphabetically we only got
to see from= and html= params which didn't contain enough information to
know which email it is. This change converts the body of mailgun request
back to a map so we can see all params for the failed request.