Add `assert_email_delivered_matches/1` to test functions
Right now if one wanted to make pretty comprehensive assertions on a
delivered email, you need to either match the exact email with
`assert_email_delivered/1` or use values or regexes with
`assert_email_delivered_with/1`. The problem with those is that it
doesn't allow you to bind a variable to the email that was delivered,
meaning you need to do all your assertions on that one email in that one
function.

This becomes especially problematic with things like testing the body of
an HTML email, as your option is either an exact match (which is
brittle) or a single regex (which is exceptionally complicated). By
allowing the user to use the match operator to bind variables and then
test those variables after the `assert_email_delivered_matches/1` call,
it can make for much simpler and less brittle tests.

It's a very small implementation - basically all it does is hide the
implementation detail of how the TestAdapter works, which I think is
enough to warrant its existence.

I thought about changing something like `assert_email_delivered/1` to
allow for binding variables in this way, but that would change the
semantics of that function too significantly, so I went with introducing
a new function.