WIP: Implement deliver_now! and deliver_later!
We start the implementation of deliver_now! and deliver_later! in
`Bamboo.Mailer`.

- `deliver_now` now returns `{:ok, email}` + response if requested or `{:error, error}`
- `deliver_now!` returns email or raises errors

One strange behavior is if `to, cc, bcc` are all empty lists. We don't
deliver, we just return `{:ok, email}` (even if `response: true`).

As we were working on that, we decide that we can introduce a
`deliver_later!`. But this one is also strange:

- `deliver_later` now returns `{:ok, email_to_send}` or `{:error,
  error}` if there was an error in the validations we perform. But if
  one of the adapters has an error in deliver later, then they might
  just have to raise.
- `deliver_later!` returns `email` and raises if there are errors.