Update TaskSupervisorStrategy to raise errors
What changed?
============

Since `adapter.deliver` should no longer raise errors, we now raise them
from the supervisor strategy. This keeps the behavior of the
TaskSupervisorStrategy the same as it was before. It'll raise and kill
the child process that was spawned to send the email.

We add a test for this, but I don't like that the output of the error is
printed out. I tried using `ExUnit.CaptureIO.capture_io(:stderr, fun)`
but it wasn't capturing it. I'll have to take another look at that. But
for now, we can keep making progress.