Fix Email.address() typespec (#570)
What changed?
============

In https://github.com/thoughtbot/bamboo/pull/386, evuez pointed out that
the `Bamboo.Email.address()` typespec shows `String.t() | {String.t(),
String.t()}`. But no adapter actually supports the `String.t()` type.

Given that the code was introduced [before the
typespecs](https://github.com/thoughtbot/bamboo/pull/150) (and thus we
can treat the code as the source of truth when in doubt), and since
`Bamboo.Email.get_address/1`'s explicit intent is to abstract the
internal representation of email addresses in case Bamboo changes how it
stores them:

> Gets just the email address from a normalized email address.
> Normalized email addresses are 2 item tuples {name, address}. This
gets the address part of the tuple. Use this instead of calling
elem(address, 1) so that if Bamboo changes how email addresses are
represented your code will still work

I think the correct change here is to remove the `String.t()` portion of
the typespec rather than update the code to handle that typespec. In
other words, I think we made a typo when introducing the typespec. So
this PR would be merged instead of #386.