Optionally specify layout in `use` call
What changed?
============

We allow the specifying of layouts when calling `use Bamboo.Template`.
That way, if someone is always using the same view and template, they
can specify it once at the top level of their `Email` module.

While doing this, we also relax the constraint that a `view` must be
specified at the `use Bamboo.Template` level. It is now possible to
simply rely on `put_view/2` when using emails.

Since that's the case, we move the error warning of a missing view
further down the call stack. Instead of throwing an error when the
`view` isn't passed as an option, we throw an error right before we
render the template.

Note: doing this created a nicer seam for the `render_email` private
function. We moved it to `Bamboo.Template` and now we call `Bamboo.View`
only for the final `render(email, template)`. That makes the separation
of concerns better: all manipulation of assigns, etc. happens in
`Bamboo.Template` and the rendering of an email and a template (as well
as the compiling of the templates) happens in `Bamboo.View`.