Import ExUnit.Assertions in Bamboo.Test.assert_delivered_email_matches
Requiring it doesn't make certain functions (like `flunk`) which are
used in Bamboo.Test available and compilation of
assert_delivered_email_matches fails when it's used outside of an ExUnit
test module. It works fine when the macro is used in an ExUnit test
because ExUnit.Assertions are imported there.