Introduce Bamboo.Template and Bamboo.View (#576)
What changed?
=============

We introduce two new modules `Bamboo.Template` and `Bamboo.View` so that
Bamboo users can use HTML and text templates without having to rely on
Phoenix (not all Bamboo users use Phoenix).

The implementation is based heavily on Phoenix's Template and View
modules. So a huge thanks to them for all the hard work.

Example usage:

```elixir
defmodule MyApp.EmailView do
  use Bamboo.View, path: "path/to/this/view/templates"
end

defmodule MyApp.Email do
  use Bamboo.Template, view: MyApp.EmailView

  def welcome_email do
    new_email()
    |> put_layout({MyApp.EmailLayout, :email})
    |> render(:welcome_email)
  end
end
```

The goal is for `Bamboo.Template` to coexist with `Bamboo.Phoenix`, and
`Bamboo.Phoenix` can be moved to a separate package in the future.

Difference from Bamboo.Phoenix
-----------------------------

Unlike `Bamboo.Phoenix`, `Bamboo.Template` allows for defining the
view with `put_view/2` for cases when we don't have to rely on the view
defined in `use Bamboo.Template, view: EmailView`.

We can also specify the layouts when calling `use Bamboo.Template`.
That way, if someone is always using the same view and template, they
can specify it once at the top level of their `Email` module.

Note on separation of concerns: all manipulation of assigns, etc. happens in
`Bamboo.Template` and the rendering of an email and a template (as well
as the compiling of the templates) happens in `Bamboo.View`.

Difference from Phoenix Views
-----------------------------

When using `Bamboo.View` we have to specify the relative path (relative
to the root of the project) where the templates for that view will live 
(instead of the root path as is done in `Phoenix.View`).

We could do fancier things where we auto detect the location based on
the module name, but that seems a little magical and perhaps unnecessary
for now. We can always add that if we need it. For now, we can specify a
view, and set the path for its templates.

Note on alternate implementation
--------------------------------

The first pass at doing this used`EEx.function_from_file` instead
`EEx.compile_file` and unquoting the compiled file as we define the
function. But I ran into issues with the `function_from_file` approach
because assigns needed to be keyword lists instead of maps. At that
point, I looked at the approach `Phoenix.Template` uses, and decided to
follow their approach of using `EEx.compile_file`.

See [Phoenix.Template's approach]:
https://github.com/phoenixframework/phoenix/blob/e05af0ad5527c9a192de122f3e43d7ed4c7a4c9f/lib/phoenix/template.ex#L346-L365

Note on updated Bamboo.Phoenix tests
------------------------------------

Even though there were no changes to `Bamboo.Phoenix`, there were a
couple of changes to `bamboo/phoenix_test`. That happened because we
tried to reuse as many as the test templates as possible.

We did separate a `PhoenixLayoutView` since the layouts render their
contents differently. `Bamboo.Phoenix` uses a version of Phoenix that
still uses `render @view_module, @view_template, assigns` in the layout.
But `Bamboo.View` is using `@inner_content`, like more recent versions
of Phoenix.

We also have to include the `function_in_view/0` function in
`bamboo/phoenix_test`'s `EmailView` since the view compiles a new
template that calls that function.