Update Mailgun.deliver to return ok/error tuples (#573)
What changed?
==============

Updates Mailgin.deliver function to return ok/error tuples to abide by
the new api.

Note on raising configuration errors
-----------------------------------

We do not change the raising of errors when they deal with configuration
options. Those errors are categorically different. We want to return
ok/error tuples when there's an error building/delivering email. If
a user of Bamboo is missing a configuration variable, it seems important
to raise an error.