Extract Bamboo.Phoenix (#581)
What changed?
=============

We extract `Bamboo.Phoenix` to a standalone library.

That makes it easier for those who want to use Bamboo but don't want to
depend on Phoenix. Prior to this commit, the Phoenix dependency was
optional. But now, those who want to use `Phoenix`'s rendering engine
can use [`bamboo_phoenix`].

[`bamboo_phoenix`]: https://github.com/thoughtbot/bamboo_phoenix