Improve interceptor/Email.block docs
What changed?
============

We improve the interceptor and Email.block docs.

In particular, we update the module doc in the Interceptor behaviour to
use the `Email.block` function rather than `Email.intercept` (which was
the function name in a previous iteration).

We also add docs to the `def block` function in `Bamboo.Email` so that
it has its own documentation.