Accept timeout option in test helpers (#605)
What changed?
============

We update all test helpers in `Bamboo.Test` to accept a `:timeout`
option. Once we get the timeout, we forward it to ExUnit's
`assert_receive` as the second argument.

Refutation helpers will use the `timeout` option passed in (if any), and
then fallback to 100ms (if not in shared mode) or the configured
`refute_timeout` if on `shared_mode`.