Set default timeout in test helpers to 100 (#607)
Commit aa8c458 introduced the ability to accept a timeout option in test
helpers. To consolidate logic, it extracted a `get_timeout/1` function,
but it incorrectly set the default value to 1000 ms. This commit
corrects that by setting it to 100 ms.

We also take the opportunity to use the `get_timeout/1` helper from the
`refute_timeout/1` function, thus we consolidate all fetching of the
`:timeout` option to the `get_timeout/1` helper.