Better local preview for text-only emails (#622)
When viewing an email message in the Bamboo.LocalAdapter's SentEmailViewerPlug, an iframe for the HTML preview is always shown and always configured to take up the whole available screen height.

This is nice in most cases. But for some internal email notifications we are happy with only the text variant. For these messages, the HTML preview is always empty and the real content is only available after scrolling down. 

This commit makes the local preview more compact for messages that have no HTML body.

It also adds color to the headings "HTML body" and "Text body" so they're more clearly visible. 

Co-authored-by: Raffael Krebs <raffael.krebs@protonmail.ch>