Add arbitrary header overrides for attachments
This allows bamboo adapters to be accept overriding headers for special
use cases. Updates to those adapters would be required to use this
functionality.

For e.g., in the specific use-case of adding inlined attachments, some
special case headers are required. Now it's possible to add
`X-Attachment-Id` and override the Content-Disposition header with
`inline` to permit the proper display of inlined content, without an
email client showing them as explicit attachments for the user to
download.